Add api mapping.

+++ server/webportal/src/main/java/ru/kolaer/server/webportal/mvc/model/dao/DAO.java
@@ -7,5 +7,5 @@ import java.util.List;
  */
 public interface Dao<T> {
     List<T> findAll();
-    T findByID(Long id);
+    T findByID(short id);
 }
